+++ packages/@react-native-windows/cli/src/e2etest/initWindows.test.ts
@@ -1,0 +1,59 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ * @format
+ */
+
+import {commanderNameToOptionName} from '@react-native-windows/telemetry';
+
+import {
+  initOptions,
+  InitOptions,
+} from '../commands/initWindows/initWindowsOptions';
+
+function validateOptionName(
+  name: string,
+  optionName: keyof InitOptions,
+): boolean {
+  // Do not add a default case here. Every item must explicitly return true
+  switch (optionName) {
+    case 'logging':
+    case 'template':
+    case 'name':
+    case 'namespace':
+    case 'overwrite':
+    case 'telemetry':
+      return true;
+  }
+  throw new Error(
+    `Unable to find ${optionName} to match '${name}' in InitOptions.`,
+  );
+}
+
+test('initOptions - validate options', () => {
+  for (const commandOption of initOptions) {
+    // Validate names
+    expect(commandOption.name).not.toBeNull();
+    expect(commandOption.name.startsWith('--')).toBe(true);
+    expect(commandOption.name).toBe(commandOption.name.trim());
+
+    // Validate defaults
+    if (
+      !commandOption.name.endsWith(' [string]') &&
+      !commandOption.name.endsWith(' [number]')
+    ) {
+      // Commander ignores defaults for flags, so leave undefined to prevent confusion
+      expect(commandOption.default).toBeUndefined();
+    }
+
+    // Validate description
+    expect(commandOption.description).not.toBeNull();
+    expect(commandOption.description!).toBe(commandOption.description!.trim());
+
+    // Validate all command options are present in InitOptions
+    const optionName = commanderNameToOptionName(commandOption.name);
+    expect(
+      validateOptionName(commandOption.name, optionName as keyof InitOptions),
+    ).toBe(true);
+  }
+});