+++ packages/playground/windows/playground-composition/Playground-Composition.cpp
@@ -60,7 +60,7 @@ struct CustomComponent : winrt::implements<CustomComponent, winrt::IInspectable>
     m_visual.Brush(m_compContext.CreateColorBrush(winrt::Windows::UI::Colors::White()));
 
     auto compositor =
-        winrt::Microsoft::ReactNative::Composition::CompositionContextHelper::InnerCompositor(m_compContext);
+        winrt::Microsoft::ReactNative::Composition::WindowsCompositionContextHelper::InnerCompositor(m_compContext);
 
     m_spotlight = compositor.CreateSpotLight();
     m_spotlight.InnerConeAngleInDegrees(50.0f);
@@ -71,9 +71,9 @@ struct CustomComponent : winrt::implements<CustomComponent, winrt::IInspectable>
     m_spotlight.LinearAttenuation(0.253f);
     m_spotlight.QuadraticAttenuation(0.58f);
     m_spotlight.CoordinateSpace(
-        winrt::Microsoft::ReactNative::Composition::CompositionContextHelper::InnerVisual(m_visual));
+        winrt::Microsoft::ReactNative::Composition::WindowsCompositionContextHelper::InnerVisual(m_visual));
     m_spotlight.Targets().Add(
-        winrt::Microsoft::ReactNative::Composition::CompositionContextHelper::InnerVisual(m_visual));
+        winrt::Microsoft::ReactNative::Composition::WindowsCompositionContextHelper::InnerVisual(m_visual));
 
     auto animation = compositor.CreateVector3KeyFrameAnimation();
     auto easeIn = compositor.CreateCubicBezierEasingFunction({0.5f, 0.0f}, {1.0f, 1.0f});
@@ -211,7 +211,7 @@ struct WindowData {
   WindowData(const winrt::Microsoft::ReactNative::CompositionHwndHost &compHost) : m_CompositionHwndHost(compHost) {
     winrt::Microsoft::ReactNative::Composition::CompositionUIService::SetCompositionContext(
         InstanceSettings().Properties(),
-        winrt::Microsoft::ReactNative::Composition::CompositionContextHelper::CreateContext(g_compositor));
+        winrt::Microsoft::ReactNative::Composition::WindowsCompositionContextHelper::CreateContext(g_compositor));
   }
 
   static WindowData *GetFromWindow(HWND hwnd) {