+++ vnext/Microsoft.ReactNative/Fabric/platform/react/renderer/components/view/HostPlatformViewProps.h
@@ -5,6 +5,7 @@
 
 #include <react/renderer/components/view/BaseViewProps.h>
 #include <react/renderer/core/PropsParserContext.h>
+#include "KeyEvent.h"
 #include "WindowsViewEvents.h"
 
 namespace facebook::react {
@@ -26,7 +27,7 @@ class HostPlatformViewProps : public BaseViewProps {
 
   // std::optional<std::string> overflowAnchor{};
   // std::optional<std::string> tooltip{};
-  // std::vector<HandledKeyEvent> keyDownEvents{};
-  // std::vector<HandledKeyEvent> keyUpEvents{};
+  std::vector<HandledKeyEvent> keyDownEvents{};
+  std::vector<HandledKeyEvent> keyUpEvents{};
 };
 } // namespace facebook::react