+++ vnext/CHANGELOG.md
@@ -1,8 +1,30 @@
 # Change Log - react-native-windows
 
-This log was last generated on Thu, 21 Sep 2023 05:13:49 GMT and should not be manually modified.
+This log was last generated on Tue, 26 Sep 2023 05:13:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0-canary.709
+
+Tue, 26 Sep 2023 05:13:15 GMT
+
+### Changes
+
+- Enable fabric usage against lifted comp, or system comp (30809111+acoates-ms@users.noreply.github.com)
+- Add new architecture `cpp-app` template for `init-windows` (jthysell@microsoft.com)
+- Bump @react-native-windows/cli to v0.0.0-canary.189
+
+## 0.0.0-canary.708
+
+Fri, 22 Sep 2023 05:13:32 GMT
+
+### Changes
+
+- [Fabric] Add support for native key event handling (30809111+acoates-ms@users.noreply.github.com)
+- add fabric textinput's cursorColor (tatianakapos@microsoft.com)
+- Create new CLI init-windows command (jthysell@microsoft.com)
+- Bump @react-native-windows/cli to v0.0.0-canary.188
+- Bump @react-native-windows/codegen to v0.0.0-canary.70
 
 ## 0.0.0-canary.707